ESSAY: Knowledge sharing within the New Zealand Government Common Web Platform: A knowledge management approach using communities of practice
This essay was written prior to the release of the Common Web Platform and my later joining SilverStripe as a Community professional. After research and applying community theories, many of the predictions highlighted in this critique have indeed proved true. This led to follow up work and guidance around the ideas of open source code, knowledge sharing and the need for communities of practice engagement to gain the strategic benefits of CWP.